+++ src/qbssettings.ts
@@ -49,6 +49,7 @@ const DEFAULT_COMMAND_ECHO_MODE = QbsCommandEchoMode.Summary;
 const DEFAULT_SHOW_DISABLED_PROJECT_ITEMS = true;
 const DEFAULT_CLEAR_OUTPUT_BEFORE_OPERATION = false;
 const DEFAULT_SAVE_BEFORE_BUILD = true;
+const DEFAULT_AUTO_RESOLVE = true;
 const DEFAULT_CONFIGURATIONS_FILE_PATH = `${SOURCE_DIR_PATTERN}/.vscode/qbs-configurations.json`;
 
 export enum QbsSettingsEvent {
@@ -208,6 +209,10 @@ export class QbsSettings implements vscode.Disposable {
 
     saveBeforeBuild(): boolean {
         return this._settings.get<boolean>('saveBeforeBuild', DEFAULT_SAVE_BEFORE_BUILD);
+    }
+
+    autoResolve(): boolean {
+        return this._settings.get<boolean>('autoResolve', DEFAULT_AUTO_RESOLVE);
     }
 
     /**